Create PokéShop Database and Stats
The goal of this project was to create my own store. The store needed to sell one type of item. I chose to make a PokéMart specializing in items found in the main series games. I used real items, prices, and badge requirements and made up any profit margins. The project required one table for all the items in the store, and columns for any data needed. Once stored, I used select statements to order the items by price and show a statistic about the items.