Create Marvel Movies and Actors
For this project I was tasked with creating a table including a set of famous people.  I chose to use Marvel actors as there are a decent amount of them to add to the table. I then created additional tables about  Marvel movies and joined it with the actors to create a list of which actors were in which movies.